UserRepository を作り，コードを移した
